feat: add debug and main function

+++ src/org/lessons/java/shop/Prodotto.java
@@ -5,17 +5,17 @@ public class Prodotto {
     int codice;
     String nome;
     String descrizione;
-    float prezzo;
+    double prezzo;
     int iva;
 
 
-    public Prodotto(String nome, String descrizione, float prezzo, int iva){
+    public Prodotto(String nome, String descrizione, double prezzo, int iva){
         
         this.nome = nome;
         this.descrizione = descrizione;
         this.prezzo = prezzo;
         this.iva = iva;
-        
+
     }
 
 }
